+++ CHANGELOG.md
@@ -7,7 +7,9 @@ _See also: [the official LibGDX changelog](https://github.com/libgdx/libgdx/blob
 - **[UPDATE]** Updated to Kotlin Coroutines 1.4.3.
 - **[FEATURE]** (`ktx-tiled`) Added `isEmpty` and `isNotEmpty` extension function for `MapLayers` and `MapObjects` collections.
 - **[FEATURE]** (`ktx-tiled`) Added `forEachLayer` extension function for `TiledMap` to iterate over an exact type 
-  of `MapLayer` instances of a specific map
+of `MapLayer` instances of a specific map.
+- **[FIX]** (`ktx-vis`) `visScrollPane` and `visSplitPane` now define the correct types for `this` in theis `init` lambdas,
+adding support for Scene2D and VisUI DSL.
 
 #### 1.9.14-b1