+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 _See also: [the official LibGDX changelog](https://github.com/libgdx/libgdx/blob/master/CHANGES)._
 
-#### 1.9.14-SNAPSHOT
+#### 1.9.14-b2
 
 - **[UPDATE]** Updated to Ashley 1.7.4.
 - **[UPDATE]** Updated to Kotlin 1.4.32.